Move more code into includes and improve use cases for our handle_form_submission function
In order to make the code easier to read and understand, I'm moving all of the similar functionality into their own files. This allows us to view the similarities in files easier, so encourages me to make more util functions.

This follows the philosophy that "Shorter code = more manageable" and "segregated code = easier to understand"