Improve security by checking for ABSPATH
Now including the if ( ! defined( 'ABSPATH' ) ) exit; statement at the top of our PHP files. This statement checks if the ABSPATH constant is defined, which would indicate that the file is being accessed through the WordPress environment. If ABSPATH is not defined, the script execution is stopped using exit, which helps to prevent direct access to the file outside of WP, enhancing the security of the plugin.